Explicity tell rivetize what props are being used by the component.
We discovered that the props of 'rivetized' components were no longer being shown in the style guide. It was determined that rivetize wasn't able to infer the props for the component it was modifying.  We made those types explicit in order that the styleguide could properly inspect the emitted components.