feat: ESIMW-2399: Fix Alert accessibility issue
The `Alert` component has an optional title prop. If this is provided, a title header tag is created with a generated id and `aria-labelledby` references this id in the parent. However, this prop was being added to the parent even when there was no title, creating a reference to an id that was not in the dom.