fix: 462 - Fix Header.Avatar display
The `Header.Avatar` and `Header.Search` components need to be rendered inside the `BaseHeaderNavigation`'s `nav` component, but outside of the navigation `ul`, which was not happening previously.

This change allows the search and avatar components to be added as children to `BaseHeaderNavigation` and the various types of components are placed in their respective locations.