feat: update to Rivet 2.x (#503)
Update the library to use Rivet 2.

BREAKING CHANGE: Rivet-react 2.0.0 is a complete rewrite of the library using Rivet 2.x. It moves the dependencies for react and react-dom to peer dependencies which allow the version of React used to be maintained by application developers as long as it meets the minimum requirements. react-icons and react-stickers are now included in the distribution package, meaning imports or script tags which reference those dependencies' .js files should be removed from client applications.

---------

Co-authored-by: ripandey <ripandey@iu.edu>
Co-authored-by: Philip Berg <pcberg@iu.edu>
Co-authored-by: Rishi Pandey <96927993+ri-pandey@users.noreply.github.com>
Co-authored-by: johglove <johglove@iu.edu>
Co-authored-by: johglove <johglove@IU-UITS-ESVM035.ads.iu.edu>
Co-authored-by: pcberg <pcberg@users.noreply.github.com>
Co-authored-by: johglove <johglove@IU-UITS-230078.ads.iu.edu>
Co-authored-by: johglove <54720060+johglove@users.noreply.github.com>
Co-authored-by: Andy Hill <athill@iu.edu>
Co-authored-by: ripandey <ripandey299@gmail.com>
Co-authored-by: vakpatel <vakpatel@iu.edu>
Co-authored-by: Vaibhavi Patel <pvaibhavi15@gmail.com>